modules: add client certificate validity time tracking to ssl_cert
This introduces `starttime` and `endtime` members to the ssl_cert class, which
correspond to the notBefore and notAfter attributes in client certificates,
respectively. This information is only exposed through the SSLINFO command,
though other components of the ircd could adapted to make use of this
information in the future.

This commit is incomplete, as it only extends the OpenSSL module to
populate these fields correctly; the GnuTLS and mbedTLS modules have not
been updated. Additionally, the module ABI has not been bumped, which may be
required given the change in the ssl_cert class.